fix:[SZ]login-google-facebook

+++ database/migrations/2024_04_14_020416_create_user_reviews_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('gambar');
             $table->foreignId('pengguna_id');
-            $table->double('penilaian');
+            $table->integer('penilaian');
             $table->string('komentar');
             $table->timestamps();
         });
